Eample updates

+++ autolens/autofit/non_linear.py
@@ -286,11 +286,14 @@ class MultiNest(NonLinearOptimizer):
         self.save_model_info()
 
         class Fitness(object):
-            def __init__(self, instance_from_physical_vector, _constant):
+
+            def __init__(self, instance_from_physical_vector, _constant, output_results):
                 self.result = None
                 self.instance_from_physical_vector = instance_from_physical_vector
                 self.constant = _constant
                 self.max_likelihood = -np.inf
+                self.output_results = output_results
+                self.accepted_samples = 0
 
             def __call__(self, cube, ndim, nparams, lnew):
 
@@ -298,17 +301,26 @@ class MultiNest(NonLinearOptimizer):
                 instance += self.constant
 
                 try:
-                    _likelihood = analysis.fit(instance)
+                    likelihood = analysis.fit(instance)
                 except exc.InversionException or exc.RayTracingException:
-                    _likelihood = -np.inf
+                    likelihood = -np.inf
 
                 # TODO: Use multinest to provide best model
 
-                if _likelihood > self.max_likelihood:
-                    self.max_likelihood = _likelihood
-                    self.result = Result(instance, _likelihood)
+                if likelihood > self.max_likelihood:
 
-                return _likelihood
+                    # TODO : make the 10 below a config file param e.g. output_results_every_accepted_samples
+
+                    self.accepted_samples += 1
+
+                    if self.accepted_samples == 10:
+                        self.accepted_samples = 0
+                        self.output_results(during_analysis=True)
+
+                    self.max_likelihood = likelihood
+                    self.result = Result(instance, likelihood)
+
+                return likelihood
 
         # noinspection PyUnusedLocal
         def prior(cube, ndim, nparams):
@@ -320,7 +332,7 @@ class MultiNest(NonLinearOptimizer):
 
             return cube
 
-        fitness_function = Fitness(self.variable.instance_from_physical_vector, self.constant)
+        fitness_function = Fitness(self.variable.instance_from_physical_vector, self.constant, self.output_results)
 
         logger.info("Running MultiNest...")
         self.run(fitness_function.__call__, prior, self.variable.total_parameters,
@@ -335,8 +347,8 @@ class MultiNest(NonLinearOptimizer):
                  log_zero=self.log_zero, max_iter=self.max_iter, init_MPI=self.init_MPI)
         logger.info("MultiNest complete")
 
+        self.output_results(during_analysis=False)
         self.output_pdf_plots()
-        self.output_results()
 
         constant = self.most_likely_instance_from_summary()
         constant += self.constant
@@ -501,46 +513,50 @@ class MultiNest(NonLinearOptimizer):
         except np.linalg.LinAlgError:
             pass
 
-    def output_results(self):
+    def output_results(self, during_analysis=False):
 
-        with open(self.file_results, 'w') as results:
+        if os.path.isfile(self.file_summary):
 
-            max_likelihood = self.max_likelihood_from_summary()
+            with open(self.file_results, 'w') as results:
 
-            results.write('Most likely model, Likelihood = ' + str(max_likelihood) + '\n')
-            results.write('\n')
+                max_likelihood = self.max_likelihood_from_summary()
 
-            most_likely = self.most_likely_from_summary()
+                results.write('Most likely model, Likelihood = ' + str(max_likelihood) + '\n')
+                results.write('\n')
 
-            for i in range(self.variable.total_parameters):
-                line = self.param_names[i]
-                line += ' ' * (50 - len(line)) + str(most_likely[i])
-                results.write(line + '\n')
+                most_likely = self.most_likely_from_summary()
 
-            most_probable = self.most_probable_from_summary()
+                for i in range(self.variable.total_parameters):
+                    line = self.param_names[i]
+                    line += ' ' * (50 - len(line)) + str(most_likely[i])
+                    results.write(line + '\n')
 
-            lower_limit = self.model_at_lower_sigma_limit(sigma_limit=3.0)
-            upper_limit = self.model_at_upper_sigma_limit(sigma_limit=3.0)
+                if during_analysis is False:
 
-            results.write('\n')
-            results.write('Most probable model (3 sigma limits)' + '\n')
-            results.write('\n')
+                    most_probable = self.most_probable_from_summary()
 
-            for i in range(self.variable.total_parameters):
-                line = self.param_names[i]
-                line += ' ' * (50 - len(line)) + str(most_probable[i]) + ' (' + str(lower_limit[i]) + ', ' + str(
-                    upper_limit[i]) + ')'
-                results.write(line + '\n')
+                    lower_limit = self.model_at_lower_sigma_limit(sigma_limit=3.0)
+                    upper_limit = self.model_at_upper_sigma_limit(sigma_limit=3.0)
 
-            lower_limit = self.model_at_lower_sigma_limit(sigma_limit=1.0)
-            upper_limit = self.model_at_upper_sigma_limit(sigma_limit=1.0)
+                    results.write('\n')
+                    results.write('Most probable model (3 sigma limits)' + '\n')
+                    results.write('\n')
 
-            results.write('\n')
-            results.write('Most probable model (1 sigma limits)' + '\n')
-            results.write('\n')
+                    for i in range(self.variable.total_parameters):
+                        line = self.param_names[i]
+                        line += ' ' * (50 - len(line)) + str(most_probable[i]) + ' (' + str(lower_limit[i]) + ', ' + str(
+                            upper_limit[i]) + ')'
+                        results.write(line + '\n')
 
-            for i in range(self.variable.total_parameters):
-                line = self.param_names[i]
-                line += ' ' * (50 - len(line)) + str(most_probable[i]) + ' (' + str(lower_limit[i]) + ', ' + str(
-                    upper_limit[i]) + ')'
-                results.write(line + '\n')
+                    lower_limit = self.model_at_lower_sigma_limit(sigma_limit=1.0)
+                    upper_limit = self.model_at_upper_sigma_limit(sigma_limit=1.0)
+
+                    results.write('\n')
+                    results.write('Most probable model (1 sigma limits)' + '\n')
+                    results.write('\n')
+
+                    for i in range(self.variable.total_parameters):
+                        line = self.param_names[i]
+                        line += ' ' * (50 - len(line)) + str(most_probable[i]) + ' (' + str(lower_limit[i]) + ', ' + str(
+                            upper_limit[i]) + ')'
+                        results.write(line + '\n')
